Big cleanup
Use the auto-generated types from the schema instead of implementing
them
