+++ packages/schema/src/types.ts
@@ -31,6 +31,55 @@ export * from './_interface/project/layers/openeoTileLayer';
 // Processing
 export * from './processing/_generated/exportProcessingSchema';
 
+// Symbology grammar
+export * from './_interface/project/symbology';
+
+import type { IGrammarLayer } from './_interface/project/symbology';
+// TODO Move into symbology folder
+
+export type ICompareOp = '>' | '<' | '>=' | '<=' | '!=';
+
+/** Full RGBA color channels. Vector layers use fill/stroke/circle channels; pixel channels for raster/KDE. */
+export type RGBAChannel =
+  | 'fill-color'
+  | 'stroke-color'
+  | 'circle-fill-color'
+  | 'circle-stroke-color'
+  | 'pixel-color';
+
+/** Integer 0–255 sub-channels — compiler assembles these into a ['color', r, g, b, a] expression. */
+export type UInt8Channel =
+  | 'fill-red'
+  | 'fill-green'
+  | 'fill-blue'
+  | 'pixel-red'
+  | 'pixel-green'
+  | 'pixel-blue';
+
+/** Float 0–1 alpha sub-channels. */
+export type UNormChannel = 'fill-alpha' | 'pixel-alpha';
+
+/**
+ * Virtual channels expanded at compile time.
+ * pixel-rgb: colorRamp/constant → R+G+B components; pair with pixel-alpha for independent alpha.
+ */
+export type VirtualChannel = 'pixel-rgb';
+
+/** Unbounded positive float channels (lengths, widths, radii). */
+export type PosFloatChannel =
+  | 'stroke-width'
+  | 'circle-radius'
+  | 'circle-stroke-width';
+
+export interface IGrammarSymbologyState {
+  /**
+   * Ordered list of independent rendering layers sharing the same source.
+   * Each layer produces one renderer layer (VectorLayer, HeatmapLayer, etc.).
+   * Layers are rendered in order (first = bottom).
+   */
+  layers: IGrammarLayer[];
+}
+
 // exportLayer
 export * from './_interface/export/exportGeoJson';
 export * from './_interface/export/exportGeoTiff';
